+++ bin/firebase_to_local.py
@@ -8,7 +8,7 @@ script_dir = os.path.dirname( __file__ )
 mymodule_dir = os.path.join( script_dir, '..' )
 sys.path.append( mymodule_dir )
 
-from nap import *
+from NAP.nap import *
 
 try:
     sys.path.append(dirname('libs/dreem/dreem'))
